+++ app/src/main/java/com/example/workouttracker/ui/nutrition/FridgeProduct.kt
@@ -2,9 +2,9 @@ package com.example.workouttracker.ui.nutrition
 
 data class FridgeProduct(
     val name: String,
-    val calories100: Int,
-    val protein100: Int,
-    val fats100: Int,
-    val carbs100: Int,
+    val calories100: Float,
+    val protein100: Float,
+    val fats100: Float,
+    val carbs100: Float,
     val availableGrams: Int? = null
 )
